# docs/api/openapi.yaml
openapi: 3.0.4
info:
  title: TechTrek API
  description: "API для пошаговой игры TechTrek (стартап-симуляции)."
  version: 0.0.2
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер разработки
...
paths:
  /auth/status:
    get:
      tags: 
        - auth
      summary: Проверить статус авторизации
      description: |
        Эндпоинт проверяет, авторизован ли пользователь и возвращает данные
        о нём, если сессия валидна.
      responses:
        '200':
          description: Статус авторизованного или неавторизованного пользователя.
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorized:
                    type: boolean
                  user:
                    type: object
                    nullable: true
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
        '401':
          description: Неавторизован (нет валидной сессии)
  /auth/register:
    post:
      tags: 
        - auth
      summary: Регистрация нового пользователя
      description: Создание новой учётной записи пользователя и возврат access/refresh токенов
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                type: string
                example: "Username is already in use"
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                type: string
                example: "An unexpected error occurred"
  /auth/login:
    post:
      tags: 
        - auth
      summary: Аутентификация пользователя
      description: Авторизация по логину и паролю. Возвращает access и refresh токены.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        '401':
          description: Неверный логин или пароль
          content:
            application/json:
              schema:
                type: string
                example: "Invalid credentials"
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                type: string
                example: "An unexpected error occurred"
...
components:
  schemas:
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        username:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT